Rename module vmm_ext to vmm
